# Labor.API/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Labor.API, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Cart": {
      "get": {
        "tags": [
          "Cart"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "sessionId",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/CartSummaryDtoApiResponse"
            }
          }
        }
      }
    },
    "/api/Cart/add": {
      "post": {
        "tags": [
          "Cart"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AddToCartDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/AddToCartResponseDtoApiResponse"
            }
          }
        }
      }
    },
    "/api/Cart/{cartId}": {
      "put": {
        "tags": [
          "Cart"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "cartId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateCartItemDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Cart"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "cartId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
...
    "/api/Cart/count": {
      "get": {
        "tags": [
          "Cart"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "sessionId",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Int32ApiResponse"
            }
          }
        }
      }
    },
    "/api/Cart/checkout": {
      "post": {
        "tags": [
          "Cart"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CheckoutRequestDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/CheckoutResponseDtoApiResponse"
            }
          }
        }
      }
    },
    "/api/Ivr/exotel/digit": {
      "post": {
        "tags": [
          "Ivr"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "Digits",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "CustomField",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "CallSid",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Ivr/mock-response": {
      "post": {
        "tags": [
          "Ivr"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/MockIvrRequestDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "AddToCartDto": {
      "required": [
        "laborId",
        "requiredHours"
      ],
      "type": "object",
      "properties": {
        "laborId": {
          "format": "int32",
          "type": "integer"
        },
        "requiredHours": {
          "format": "int32",
          "maximum": 24,
          "minimum": 1,
          "type": "integer"
        },
        "workDescription": {
          "maxLength": 1000,
          "minLength": 0,
          "type": "string"
        },
        "preferredDate": {
          "format": "date-time",
          "type": "string"
        },
        "sessionId": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "AddToCartResponseDto": {
      "type": "object",
      "properties": {
        "cartId": {
          "format": "int32",
          "type": "integer"
        },
        "sessionId": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "AddToCartResponseDtoApiResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/AddToCartResponseDto"
        },
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "additionalProperties": false
    },
...
    "ApiResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        },
        "data": { },
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "additionalProperties": false
    },
...
    "CartItemDto": {
      "type": "object",
      "properties": {
        "cartId": {
          "format": "int32",
          "type": "integer"
        },
        "laborId": {
          "format": "int32",
          "type": "integer"
        },
        "requiredHours": {
          "format": "int32",
          "type": "integer"
        },
        "dailyRate": {
          "format": "double",
          "type": "number"
        },
        "totalAmount": {
          "format": "double",
          "type": "number"
        },
        "workDescription": {
          "type": "string"
        },
        "preferredDate": {
          "format": "date-time",
          "type": "string"
        },
        "createdAt": {
          "format": "date-time",
          "type": "string"
        },
        "laborName": {
          "type": "string"
        },
        "profilePicture": {
          "type": "string"
        },
        "laborType": {
          "type": "string"
        },
        "specialization": {
          "type": "string"
        },
        "rating": {
          "format": "double",
          "type": "number"
        },
        "totalReviews": {
          "format": "int32",
          "type": "integer"
        },
        "isExpired": {
          "type": "boolean"
        },
        "isUnavailableNow": {
          "type": "boolean"
        }
      },
      "additionalProperties": false
    },
    "CartSummaryDto": {
      "type": "object",
      "properties": {
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CartItemDto"
          }
        },
        "totalItems": {
          "format": "int32",
          "type": "integer"
        },
        "totalAmount": {
          "format": "double",
          "type": "number"
        }
      },
      "additionalProperties": false
    },
    "CartSummaryDtoApiResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/CartSummaryDto"
        },
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "additionalProperties": false
    },
...
    "CheckoutRequestDto": {
      "type": "object",
      "properties": {
        "sessionId": {
          "type": "string"
        },
        "addressId": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
    "CheckoutResponseDto": {
      "type": "object",
      "properties": {
        "cartSummary": {
          "$ref": "#/definitions/CartSummaryDto"
        },
        "message": {
          "type": "string"
        },
        "requireAddress": {
          "type": "boolean"
        },
        "requirePayment": {
          "type": "boolean"
        }
      },
      "additionalProperties": false
    },
    "CheckoutResponseDtoApiResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/CheckoutResponseDto"
        },
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "additionalProperties": false
    },
...
    "Int32ApiResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "format": "int32",
          "type": "integer"
        },
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "additionalProperties": false
    },
...
    "MockIvrRequestDto": {
      "type": "object",
      "properties": {
        "laborConfirmationId": {
          "format": "int32",
          "type": "integer"
        },
        "digit": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "UpdateCartItemDto": {
      "type": "object",
      "properties": {
        "laborId": {
          "format": "int32",
          "type": "integer"
        },
        "requiredHours": {
          "format": "int32",
          "type": "integer"
        },
        "workDescription": {
          "type": "string"
        },
        "preferredDate": {
          "format": "date-time",
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
  },
  "securityDefinitions": {
    "Bearer": { }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}